Make to_str a shared function in python
