+++ jsone/render.py
@@ -2,9 +2,8 @@ from __future__ import absolute_import, print_function, unicode_literals
 
 import re
 import json as json
-from .shared import JSONTemplateError, TemplateError, DeleteMarker, string
+from .shared import JSONTemplateError, TemplateError, DeleteMarker, string, to_str
 from . import shared
-from . import builtins
 from .interpreter import ExpressionEvaluator
 from .six import viewitems
 
@@ -40,7 +39,7 @@ def interpolate(string, context):
             if isinstance(parsed, (list, dict)):
                 raise TemplateError(
                     "interpolation of '{}' produced an array or object".format(string[:offset]))
-            result.append(builtins.build({})['str'](parsed))
+            result.append(to_str(parsed))
             string = string[offset + 1:]
         else:  # found `$${`
             result.append('${')
